+++ ZeldaClone/src/States/PauseState.h
@@ -8,7 +8,7 @@ public:
 	PauseState();
 	~PauseState() {}
 
-	virtual void Update(const double& deltaTime) override;
+	virtual void Update(const float& deltaTime) override;
 	virtual void Render() override;
 
 	virtual bool OnEnter() override;
